+++ src/simulation/initializers.jl
@@ -12,7 +12,7 @@ function newImmunity!(imprinted_pathogen::Pathogen, matured_pathogen::Pathogen, 
     push!(class.immunities, Immunity(
         length(class.immunities) + 1, imprinted_pathogen, matured_pathogen,
         immunityStaticCoefficients(imprinted_pathogen.sequence, matured_pathogen.sequence, type),
-        type.id
+        type
     ))
 
     return class.immunities[end]
@@ -32,8 +32,8 @@ function newHost!(class::Class, population::Population, model::Model)
     propagateWeightChanges!(
         SVector{NUM_COEFFICIENTS,Float64}(
             [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0]
-            # class change, inter-population contact and migration numbers per class or
-            # population are fractions that sum to one, so no need to account for in here
+        # class change, inter-population contact and migration numbers per class or
+        # population are fractions that sum to one, so no need to account for in here
         ), length(class.hosts), class, population, model
     )